Created Baseclass for integration testing
