+++ Assemblies/Ms18.Application/appsettings.Development.json
@@ -6,6 +6,6 @@
     }
   },
   "ConnectionStrings": {
-    "DefaultConnection": "UserID=postgres;Password=postgres;Host=localhost;port=5410;Database=Maasgroep;Pooling=true"
+    "DefaultConnection": "UserID=postgres;Password=postgres;Host=localhost;port=5432;Database=Maasgroep;Pooling=true"
   }
 }
