--- Tests/Ms18.Application.Test/appsettings.Development.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "DefaultConnection": "UserID=postgres;Password=postgres;Host=localhost;port=5410;Database=Maasgroep;Pooling=true"
-  }
-}
